Check in image uploader. Only admins are allowed to upload any file. Added ACL to allow uploading of other media types than images fixed filter for arrays added eu_vat_id Added tab emails removed old $user->authorise in media getFiles removed Sql injection in order model, was only useable as admin, so the treat is low more filters for getOrderByList added different offset for years for date and ageverification userfields, maybe a parameter later
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/trunk/virtuemart@9009 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/helpers/image.php
@@ -27,17 +27,22 @@ class VmImage extends VmMediaHandler {
 			$oldFileUrl = $this->file_url;
 			$file_name = $this->uploadFile($this->file_url_folder);
 			if($file_name){
+
 				if($file_name!=$oldFileUrl && !empty($this->filename)){
 					$this->deleteFile($oldFileUrl);
 				}
 				$this->file_url = $this->file_url_folder.$file_name;
 				$this->filename = $file_name;
 
-				$oldFileUrlThumb = $this->file_url_thumb;
-				$this->file_url_thumb = $this->createThumb();
-				if($this->file_url_thumb!=$oldFileUrlThumb){
-					$this->deleteFile($oldFileUrlThumb);
+				if(!empty($this->file_url_thumb)){
+					$oldFileUrlThumb = $this->file_url_thumb;
+
+					$this->file_url_thumb = $this->createThumb();
+					if($this->file_url_thumb!=$oldFileUrlThumb){
+						$this->deleteFile($oldFileUrlThumb);
+					}
 				}
+
 			}
 		} //creating the thumbnail image
 		else if( $data['media_action'] == 'create_thumb' ){
